Level Select screen
I got the level select screen to work. When you select play on the title screen, it now updates the screen to show the different levels that can be selected. It also allows you to select them, but like with the title screen, it only works with mouse input as of the moment.