level_list, restarting, score logic
I added a level_list that, well, should be self-explanatory in what it is. The match-case in the game loop builds the level into the list at the moment the player selects the level. This way seemed easiest for allowing levels to be built differently based off of the user's level choice instead of pre-building them. Also fixed a logic error with the score timer starting at the title screen, so the score would already be higher than 0 by the time a level was selected (thankfully no-one seemed to notice that during Hunter's presentation before I did). The play, quit, and level "buttons" now have actual button logic to them, where hovering over them with the mouse draws a rectangle around them indicating it is selectable to the user. Just to begin familiarizing myself w/ the level classes, I put an intersection into our level 1 demo. Finally, added restart logic on the game over screen that resets the cars to their original positions and the score back to zero when the user presses space.