Adjustments
I adjusted some of the UI so that the road would display more centered. I also began on the format for the tutorial. There were some changes I made to the Level4.py that I didn't commit here, but it was the only way to get them to work in the match-case I made for the UI (I'll talk about it at meeting later today).