Bot cars reset
Bot cars reset when they go off-screen now. Also, a fail condition has been added if the player does not stop at the stop sign (does this by checking the velocity of the player car between a certain range before the stop sign and setting a flag if the velocity hits zero anywhere in that range).